Wall fixes, once again.

+++ src/libs/Adafruit-GFX-Library/gfxfont.h
@@ -14,8 +14,8 @@ typedef struct
     uint8_t width;         ///< Bitmap dimensions in pixels
     uint8_t height;        ///< Bitmap dimensions in pixels
     uint8_t xAdvance;      ///< Distance to advance cursor (x axis)
-    int8_t xOffset;        ///< X dist from cursor pos to UL corner
-    int8_t yOffset;        ///< Y dist from cursor pos to UL corner
+    int xOffset;           ///< X dist from cursor pos to UL corner
+    int yOffset;           ///< Y dist from cursor pos to UL corner
 } GFXglyph;
 
 /// Data stored for FONT AS A WHOLE
